+++ packages/terra-time-input/src/terra-dev-site/test/time-input/time-input/DefaultTime.test.jsx
@@ -1,12 +1,18 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import TimeInput from '../../../../TimeInput';
+import styles from './common/TimeInput.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const TimeInputDefault = () => (
-  <TimeInput
-    id="timeInput"
-    name="time-input"
-    value="12:00"
-  />
+  <div className={cx('content-wrapper')}>
+    <TimeInput
+      id="timeInput"
+      name="time-input"
+      value="12:00"
+    />
+  </div>
 );
 
 export default TimeInputDefault;