+++ packages/terra-time-input/src/terra-dev-site/test/time-input/time-input/Disabled.test.jsx
@@ -1,14 +1,20 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import TimeInput from '../../../../TimeInput';
+import styles from './common/TimeInput.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const TimeInputDefault = () => (
-  <TimeInput
-    id="timeInput"
-    name="time-input"
-    value="09:22"
-    variant="12-hour"
-    disabled
-  />
+  <div className={cx('content-wrapper')}>
+    <TimeInput
+      id="timeInput"
+      name="time-input"
+      value="09:22"
+      variant="12-hour"
+      disabled
+    />
+  </div>
 );
 
 export default TimeInputDefault;