+++ packages/terra-time-input/tests/wdio/time-input-spec.js
@@ -24,7 +24,43 @@ Terra.describeViewports('Time Input', ['medium'], () => {
   });
 
   describe('Invalid time provided', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/invalid-provided-time'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Invalid time input', () => {
     before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/invalid-time'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Invalid mobile time input', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/mobile-invalid'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Incomplete time input', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/incomplete-time'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Incomplete mobile time input', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/mobile-incomplete'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Invalid and Incomplete time input', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/invalid-incomplete-time'));
+
+    Terra.it.validatesElement();
+  });
+
+  describe('Invalid and Incomplete mobile time input', () => {
+    before(() => browser.url('/#/raw/tests/terra-time-input/time-input/time-input/mobile-invalid-incomplete'));
 
     Terra.it.validatesElement();
   });