+++ packages/terra-brand-footer/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+2.17.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 2.16.0 - (October 3, 2019)
 ------------------
 ### Changed