+++ packages/terra-date-picker/CHANGELOG.md
@@ -6,6 +6,15 @@ Unreleased
 ### Changed
 * Update Jest snapshots
 
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+
 4.15.0 - (October 3, 2019)
 ------------------
 ### Changed