+++ packages/terra-disclosure-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+4.24.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.23.0 - (October 3, 2019)
 ------------------
 ### Changed